Add files via upload
This project is a front-end replica of the popular e-commerce platform Myntra, built using HTML and CSS. It showcases a fully responsive design with a user-friendly interface, mimicking the look and feel of the original website. 
Technologies Used:
HTML5: For creating the structure and content of the pages.
CSS3: For styling, layout, and animations.
Purpose:
This project was created to enhance my skills in web design and front-end development. It demonstrates my ability to replicate complex designs using pure HTML and CSS while ensuring responsiveness and a modern aesthetic.